[#117] Fix 'betas' for positive even indexes
The problem with 'betasEven' has been fixed and it now produces
correct results. For example, 'β(2)' is now Catalan's constant.

The problem was twofold:
* In the second summand of the RHS of the equation used to derive
the recurrence ([1]), the 'ζ' function was called starting
at 'ζ(1)' and not at 'ζ(3)' as it should have been.
* The second factor in the third addend of the RHS (see below) was
giving the wrong denominator to the odd powers of 'pi'
(starting at 3).
Instead of using '[pi^3 / 2^4, pi^5 / 2^6, pi^7 / 2^8 ..]' it
was using '[pi^3 / 4, pi^5 / 6, pi^7 / 8 ..]'.

[1]: [An Euler-type formula for β(2n) and closed-form expressions for a class of zeta series](https://arxiv.org/pdf/0910.5004.pdf)
by F.M.S. Lima. See formula (12) in page 5.