[#117] Refactor 'betas' to improve numerical stability
'betas' is now more accurate, but it is still not optimal for
positive even integers. The fact that for these values the
algorith chosen depends on 'zetasOdd', which itself then depends
on an approximated infinite sum using 'zetasEven' makes it quite
unstable.
A few more tests mirroring those for 'zetas' have been added.