[#118] Add `findPrime` function
In addition, the `isPrime` function has been fixed as it
contained two errors.

- Firstly, it checked if the norm of
an Eisenstein integer was congruent to 0 modulo 3, which
is not necessary because it is supposed to be a prime
integer. This is only valid for `1 - ω`, which has been
moved into its own patter match.

- Secondly, it did not account for the possibility of an
Eisenstein integer being the product of an integer prime
and an Eisenstein unit. For example, `isPrime 11` would be
`True`, but `isPrime $ 11 * (1 + ω)` would be `False`, which
is incorrect.