Remove `Word/Int` duplicate tests for `smoothOver`
Because smooth numbers tend to grow very quickly, the tests
checking that `smoothOver` does not generate duplicates for the
`Int/Word` types overflow quickly, which creates a lot of zeroes
that `nub` then removes, leaving only one. This means the
duplicate-free list is different than the original, causing the
test to fail.

As such, the tests are only run for `Natural/Integer` now.